refactor: reorganize project structure for professional development
- move all configuration files to config/ folder for centralization
- reorganize test structure by type (unit/, e2e/, integration/)
- move development scripts to tools/ folder
- clean up root directory from 25+ files to 15 essential files
- update all package.json scripts to reference new config paths
- update husky hooks to use new config locations
- update lint-staged configuration for new file structure
- add comprehensive spell check dictionary (fixed 72 issues)
- improve project documentation with new structure guide
- maintain full functionality while achieving enterprise-grade organization
