+++ .husky/commit-msg
@@ -1,1 +1,1 @@
-npx --no -- commitlint --edit \
+npx --no -- commitlint --edit --config ./config/commitlint.config.js
