+++ config/.textlintrc.json
