+++ package.json
@@ -10,43 +10,43 @@
 		"preview": "astro preview",
 		"astro": "astro",
 		"// === Code Quality ===": "",
-		"lint": "eslint . --ext .js,.ts,.astro,.jsx,.tsx --fix",
-		"lint:text": "textlint \"src/content/**/*.{md,mdx}\" \"README.md\" \"*.md\"",
-		"lint:text:fix": "textlint --fix \"src/content/**/*.{md,mdx}\" \"README.md\" \"*.md\"",
-		"lint:spell": "cspell \"**/*.{ts,js,astro,tsx,jsx,md,mdx}\" --no-progress",
-		"lint:spell:fix": "cspell \"**/*.{ts,js,astro,tsx,jsx,md,mdx}\" --no-progress --show-suggestions",
+		"lint": "eslint . --ext .js,.ts,.astro,.jsx,.tsx --fix --config ./config/eslint.config.js",
+		"lint:text": "textlint \"src/content/**/*.{md,mdx}\" \"README.md\" \"*.md\" --config ./config/.textlintrc.json",
+		"lint:text:fix": "textlint --fix \"src/content/**/*.{md,mdx}\" \"README.md\" \"*.md\" --config ./config/.textlintrc.json",
+		"lint:spell": "cspell \"**/*.{ts,js,astro,tsx,jsx,md,mdx}\" --no-progress --config ./config/cspell.json",
+		"lint:spell:fix": "cspell \"**/*.{ts,js,astro,tsx,jsx,md,mdx}\" --no-progress --show-suggestions --config ./config/cspell.json",
 		"lint:all": "npm run lint && npm run lint:text:fix && npm run lint:text && npm run lint:spell",
-		"format": "prettier --write .",
+		"format": "prettier --write . --config ./config/.prettierrc --ignore-path ./config/.prettierignore",
 		"// === Testing - Core ===": "",
-		"test": "vitest",
-		"test:run": "vitest run",
-		"test:watch": "vitest --watch",
-		"test:ui": "vitest --ui",
-		"test:coverage": "vitest run --coverage",
+		"test": "vitest --config ./config/vitest.config.ts",
+		"test:run": "vitest run --config ./config/vitest.config.ts",
+		"test:watch": "vitest --watch --config ./config/vitest.config.ts",
+		"test:ui": "vitest --ui --config ./config/vitest.config.ts",
+		"test:coverage": "vitest run --coverage --config ./config/vitest.config.ts",
 		"// === Testing - Suites ===": "",
-		"test:critical": "vitest run test/grammar-checker.test.ts test/mdx.test.ts test/seo.test.ts",
-		"test:integration": "vitest run test/links.test.ts test/accessibility.test.ts test/admin-interface.test.ts",
-		"test:performance": "vitest run test/performance.test.ts",
-		"test:back-to-top": "vitest run back-to-top",
+		"test:critical": "vitest run test/unit/grammar-checker.test.ts test/unit/mdx.test.ts test/unit/seo.test.ts --config ./config/vitest.config.ts",
+		"test:integration": "vitest run test/e2e/links.test.ts test/e2e/accessibility.test.ts test/admin-interface.test.ts --config ./config/vitest.config.ts",
+		"test:performance": "vitest run test/performance.test.ts --config ./config/vitest.config.ts",
+		"test:back-to-top": "vitest run back-to-top --config ./config/vitest.config.ts",
 		"// === Testing - Chrome AI ===": "",
-		"ai:unit": "vitest run test/utils/summarizer.test.ts test/utils/translator.test.ts",
-		"ai:integration": "vitest run test/integration/chrome-ai-components.test.ts",
-		"ai:performance": "vitest run test/performance/chrome-ai-performance.test.ts",
-		"ai:errors": "vitest run test/error-handling/chrome-ai-errors.test.ts",
+		"ai:unit": "vitest run test/utils/summarizer.test.ts test/utils/translator.test.ts --config ./config/vitest.config.ts",
+		"ai:integration": "vitest run test/integration/chrome-ai-components.test.ts --config ./config/vitest.config.ts",
+		"ai:performance": "vitest run test/performance/chrome-ai-performance.test.ts --config ./config/vitest.config.ts",
+		"ai:errors": "vitest run test/error-handling/chrome-ai-errors.test.ts --config ./config/vitest.config.ts",
 		"ai:all": "npm run ai:unit && npm run ai:integration && npm run ai:performance && npm run ai:errors",
 		"// === Testing - Features ===": "",
-		"test:admin": "vitest run test/admin-interface.test.ts",
-		"test:grammar": "vitest run test/grammar-checker.test.ts",
-		"test:date": "vitest run test/date-filtering.test.ts",
+		"test:admin": "vitest run test/admin-interface.test.ts --config ./config/vitest.config.ts",
+		"test:grammar": "vitest run test/unit/grammar-checker.test.ts --config ./config/vitest.config.ts",
+		"test:date": "vitest run test/date-filtering.test.ts --config ./config/vitest.config.ts",
 		"// === Quality Checks ===": "",
-		"spell": "cspell \"**/*.{ts,js,astro,tsx,jsx,md,mdx}\" --no-progress",
-		"spell:check": "cspell \"src/**/*.{ts,js,astro,tsx,jsx}\" \"*.md\" --no-progress",
+		"spell": "cspell \"**/*.{ts,js,astro,tsx,jsx,md,mdx}\" --no-progress --config ./config/cspell.json",
+		"spell:check": "cspell \"src/**/*.{ts,js,astro,tsx,jsx}\" \"*.md\" --no-progress --config ./config/cspell.json",
 		"check": "npm run lint:all && npm run test:critical",
 		"check:full": "npm run lint:all && npm run test:run",
 		"// === Pre-push Checks ===": "",
 		"pre-push": "npm run check:full && npm run build && npx astro check",
 		"pre-push:fast": "npm run test:critical && npm run build",
-		"check:commit": "tsx scripts/test-grammar.ts",
+		"check:commit": "tsx tools/test-grammar.ts",
 		"// === Utilities ===": "",
 		"clean": "tsx scripts/test-clean.ts",
 		"fast": "tsx scripts/test-fast.ts",
@@ -126,14 +126,14 @@
 	},
 	"lint-staged": {
 		"*.{js,jsx,ts,tsx,astro}": [
-			"prettier --write",
-			"eslint --fix",
-			"cspell --no-progress"
+			"prettier --write --config ./config/.prettierrc --ignore-path ./config/.prettierignore",
+			"eslint --fix --config ./config/eslint.config.js",
+			"cspell --no-progress --config ./config/cspell.json"
 		],
 		"*.{md,mdx}": [
-			"prettier --write",
-			"textlint --fix",
-			"cspell --no-progress"
+			"prettier --write --config ./config/.prettierrc --ignore-path ./config/.prettierignore",
+			"textlint --fix --config ./config/.textlintrc.json",
+			"cspell --no-progress --config ./config/cspell.json"
 		]
 	}
 }
