+++ src/lib/utils.ts
@@ -1,4 +1,4 @@
-import { type ClassValue,clsx } from 'clsx'
+import { type ClassValue, clsx } from 'clsx'
 import { twMerge } from 'tailwind-merge'
 
 export function cn(...inputs: ClassValue[]) {
