+++ test/back-to-top.test.ts
@@ -1,6 +1,6 @@
 import { spawn } from 'child_process'
 import puppeteer from 'puppeteer'
-import { afterAll,beforeAll, describe, expect, it } from 'vitest'
+import { afterAll, beforeAll, describe, expect, it } from 'vitest'
 
 // Helper function to wait for server to be ready
 async function waitForServer(
