+++ test/date-filtering.test.ts
@@ -1,5 +1,5 @@
 import { ChildProcess, spawn } from 'child_process'
-import { afterAll,beforeAll, describe, expect, it } from 'vitest'
+import { afterAll, beforeAll, describe, expect, it } from 'vitest'
 
 interface Link {
 	id: number
