+++ test/e2e/accessibility.test.ts
@@ -1,7 +1,7 @@
-import { readdirSync,readFileSync } from 'fs'
+import { readdirSync, readFileSync } from 'fs'
 import matter from 'gray-matter'
 import { join } from 'path'
-import { beforeAll,describe, expect, it } from 'vitest'
+import { beforeAll, describe, expect, it } from 'vitest'
 
 interface BlogPostData {
 	title: string
