+++ tools/check-commit-grammar.ts
