+++ tools/test-clean.ts
