+++ tools/test-fast.ts
